A1 setup

+++ src/assignment1/BoardEvaluator.java
@@ -1,5 +1,5 @@
 package assignment1;
 
 public interface BoardEvaluator {
-    int evaluateBoard(Board board);
+    double evaluateBoard(Board board);
 }
